% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IplBoost.R
\name{cv.IplBoost}
\alias{cv.IplBoost}
\title{cv.IplBoost}
\usage{
cv.IplBoost(times, status, mat, lms, w, M, lambda, folds, verbose = FALSE,
  standardise = TRUE, parallel = FALSE)
}
\arguments{
\item{times}{A n-dimensional vector of survival times}

\item{status}{A n-dimensional vector of censoring indicators}

\item{mat}{A n x p matrix of covariate values}

\item{lms}{A S-dimensinal vector of landmark points to produce dynamic
predictions from}

\item{w}{A number. The "landmark interval width" or how far ahead the survival
predictions will be made}

\item{M}{A number. The number maximum of boosting iterations to perform}

\item{lambda}{A number or an S-dimensional vector. The regularisation parameter 
for the boosting algorithm.}

\item{folds}{A n-dimensional vector that assigns a number between 1 and K to 
each observations for cross-validataion.}

\item{verbose}{Boolean. Indicates whether the iteration number should be 
printed to the console.}

\item{standardise}{Boolean. Indicates if covariates should be standardised.}

\item{parallel}{Boolean. Indicates if the cross validation should be performed in parallel.
relies on the package snowfall. Cluster must be initialised before calling cv.IplBoost in the
case of this being true, see the snowfall package and ?sfInit.}
}
\value{
A list containg a (M+1)-dimensional vector of the cross validated ipl as the first element,
and a number indicating the optimal number of iterations as the second.
}
\description{
This function performs K-fold cross-valitation for \link{IplBoost},
to tune the number of iterations.
}
\examples{
# Tune the number of iterations via cross validation
cv.mod <- cv.IplBoost(times, status, design, lms=seq(0, 10, 0.1),
                      w=5, M=100, lambda=100,
                      folds=Kfold(length(times), 10))
                      
# Plot the cross-validated likelihood
plot(0:100, cv.mod$ipl.cv, "l", lwd=2, xlab="Number of iterations",
     ylab="Cross-validated ipl")
}
