\name{IplBoost-package}
\alias{IplBoost-package}
\alias{IplBoost}
\docType{package}
\title{
  IplBoost: Sliding landmark models fitted via likelihood based boosting.
}
\description{
  A more detailed description of what the package does. A length
  of about one to five lines is recommended.

  testing tex:
  $\beta$
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Simon Boge Brant

Maintainer: Your Name <your@email.com>
}
\references{
  
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
